+++ custom_components/home_status/const.py
@@ -3,5 +3,5 @@
 DOMAIN = "home_status"
 URL_BASE = "/home_status"
 JSMODULES = (
-    {"filename": "home-status-card.js", "version": "0.8.2"},
+    {"filename": "home-status-card.js", "version": "0.9.4"},
 )